requirements.txt
apache-airflow
typing-extensions
psycopg2
vk_api
